feat(lib-common): promote hat color/shape vocabulary from app-vscode
Moves HAT_COLORS / HAT_SHAPES / HAT_NON_DEFAULT_SHAPES and the HatColor /
HatShape / HatNonDefaultShape / VscodeHatStyleName types out of app-vscode
into lib-common's hatStyles.types (which previously held only the
HatStyleName stub), so non-VS-Code consumers (e.g.
@cursorless/command-visualizer) can import the vocabulary instead of
cloning it.

app-vscode's hatStyles.types.ts becomes a backward-compatible re-export
shim, so its ~10 consumers (VscodeHats, VscodeHatRenderer, getStyleName,
keyboard/*, hatAdjustments scripts, ...) are unchanged. No behavior change.
